282 287 atom positioning (#297)
* Fixed an unrelated segfault that was an obstacle.

* Fix for #282 before work can be done on #287.

* Upgrade of Point class to double precision.

* Test file for coplanar pi hydrogenation.

* Pepteditor script for coplanar pi hydrogen test.

* Test of coplanar pi atoms after hydrogenation.

* Updated tests affected by the Point precision.

* Fixed array deletion before array usage.

* Observe pi coplanarity when hydrogenating.

* Updates to tests.

* Retrying...

* Fixed infinite loop.

* Temporarily increasing limit. See #288.

* Unit test.

* Unit test enhancements.

+++ src/interpreter.cpp
@@ -1224,7 +1224,7 @@ int main(int argc, char** argv)
                 case SV_POINT:
                     if (flags & _HAS_DOT)
                     {
-                        float* ff = nullptr;
+                        double* ff = nullptr;
 
                         char* param = strchr(words[1], '.');
                         if (!param) raise_error( (std::string)"Missing member after dot.");
